+++ src/assets/lang/de/home.json
@@ -64,7 +64,7 @@
       "description": {
         "normal1": "",
         "blue": "80% RABATT ",
-        "normal2": "auf alle Premium-Pläne."
+        "normal2": "auf alle Lifetime-Pläne."
       },
       "features": ["Bis zu 10TB Speicher", "Zero-Knowledge-Verschlüsselung", "Open source und geprüft"],
       "startFrom": {
