+++ src/assets/lang/de/metatags-descriptions.json
@@ -144,5 +144,11 @@
     "id": "cloud-storage-for-education",
     "title": "Cloud Storage Discount for Education | Internxt",
     "description": "Internxt cloud storage discount for education, schools, and universities. 50% off any cloud storage subscription plans. Keep your files secure with Internxt!"
+  },
+  {
+    "id": "halloween-campaign",
+    "title": "Internxt – Private & Sichere Cloud-Speicherung",
+    "description": "Internxt Halloween-Kampagne: 80 % Rabatt auf all unsere Jahrespläne. Sichern Sie Ihre Dateien ein Leben lang mit nur einer Zahlung."
   }
+
 ]
