+++ src/assets/lang/en/banners.json
@@ -145,8 +145,8 @@
   },
   "TopBarBanner": {
     "title": {
-      "boldText": "80% OFF your next lifetime plans.",
-      "normalText": "Killer savings!",
+      "boldText": "80% off your next lifetime plan",
+      "normalText": "Hello, I want to play a game!",
       "cta": "Make your choice"
     }
   },
