+++ src/assets/lang/en/home.json
@@ -65,7 +65,7 @@
       "description": {
         "normal1": "",
         "blue": "80% off ",
-        "normal2": "all premium plans"
+        "normal2": "all lifetime plans"
       },
       "features": ["Up to 10TB storage", "Zero knowledge encryption", "Open source and audited"],
       "startFrom": {
