+++ src/assets/lang/es/home.json
@@ -65,7 +65,7 @@
       "description": {
         "normal1": "",
         "blue": "80% DE DESCUENTO ",
-        "normal2": "en todos los planes"
+        "normal2": "en los planes Lifetime"
       },
       "features": ["Hasta 10TB de almacenamiento", "Cifrado de conocimiento cero", "Código abierto y auditado"],
       "startFrom": {
