+++ src/assets/lang/es/priceCard.json
@@ -96,12 +96,13 @@
       "10TB": ["Fondos de pantalla Star Wars", "Star Wars Theme para la Web App", "Stickers de Star Wars"]
     },
      "halloweenFeatures":{
-      "title": "Internxt themed goodies:",
-      "200GB": ["Internxt x Swiss Peak backpack", "Halloween theme for Web App"],
-      "1TB": ["Internxt x Swiss Peak backpack", "Halloween theme for Web App"],
-      "2TB":  ["Internxt x Swiss Peak backpack", "Halloween theme for Web App"],
-      "5TB":  ["Internxt x Swiss Peak backpack", "Halloween theme for Web App"],
-      "10TB":  ["Internxt x Swiss Peak backpack", "Halloween theme for Web App"]
+      "title": "Regalitos de Internxt",
+      "200GB":["Mochila Internxt x Swiss Peak ", "Tema de Halloween para Web App"],
+      "1TB":  ["Mochila Internxt x Swiss Peak ", "Tema de Halloween para Web App"],
+      "2TB":  ["Mochila Internxt x Swiss Peak ", "Tema de Halloween para Web App"],
+      "5TB":  ["Mochila Internxt x Swiss Peak ", "Tema de Halloween para Web App"],
+      "10TB": ["Mochila Internxt x Swiss Peak ", "Tema de Halloween para Web App"],
+      "tooltip":"Comprando cualquier plan lifetime!"
     },
     "cheaperThan": {
       "lifetime": {
