+++ src/assets/lang/fr/banners.json
@@ -89,9 +89,9 @@
   },
   "TopBarBanner": {
     "title": {
-      "boldText": "80% de réduction sur tous les plans!",
-      "normalText": "Économies pour le Mois de la Sensibilisation à la Cybersécurité!",
-      "cta": "Réclamez l'offre"
+      "boldText": "80 % de réduction sur votre prochain forfait à vie",
+      "normalText": "Bonjour, je veux jouer à un jeu !",
+      "cta": "Oses-tu ?"
     }
   },
   "lifetimeBanner": {
