+++ src/assets/lang/fr/home.json
@@ -65,7 +65,7 @@
       "description": {
         "normal1": "",
         "blue": "80% DE RÉDUCTION ",
-        "normal2": "sur tous les plans premium."
+        "normal2": "sur tous les plans à vie"
       },
       "features": ["Jusqu'à 10TB de stockage", "Chiffrement sans connaissance", "Open source et audité"],
       "startFrom": {
