+++ src/assets/lang/fr/metatags-descriptions.json
@@ -128,5 +128,10 @@
     "id": "cloud-storage-for-education",
     "title": "Cloud Storage Discount for Education | Internxt",
     "description": "Internxt cloud storage discount for education, schools, and universities. 50% off any cloud storage subscription plans. Keep your files secure with Internxt!"
+  },
+  {
+    "id": "halloween-campaign",
+    "title": "Internxt – Stockage en nuage privé et sécurisé",
+    "description": "Campagne d'Halloween d'Internxt : 80 % de réduction sur tous nos plans annuels. Protégez vos fichiers pour la vie avec un seul paiement."
   }
 ]
