+++ src/assets/lang/it/banners.json
@@ -117,9 +117,9 @@
   },
   "TopBarBanner": {
     "title": {
-      "boldText": "80% di sconto su tutti i piani!",
-      "normalText": "Risparmi per il Mese della Consapevolezza Cibernetica!",
-      "cta": "Richiedi l'offerta"
+      "boldText": "80% di sconto sul tuo prossimo piano a vita",
+      "normalText": "Ciao, voglio giocare a un gioco!",
+      "cta": "Hai il coraggio?"
     }
   },
   "lifetimeBanner": {
