+++ src/assets/lang/it/home.json
@@ -65,7 +65,7 @@
       "description": {
         "normal1": "",
         "blue": "80% DI SCONTO ",
-        "normal2": "su tutti i piani premium."
+        "normal2": "su tutti i piani a vita"
       },
       "features": ["Fino a 10TB di storage", "Crittografia a conoscenza zero", "Open source verificato"],
       "startFrom": {
