+++ src/assets/lang/it/metatags-descriptions.json
@@ -148,5 +148,11 @@
     "id": "cloud-storage-for-education",
     "title": "Cloud Storage Discount for Education | Internxt",
     "description": "Internxt cloud storage discount for education, schools, and universities. 50% off any cloud storage subscription plans. Keep your files secure with Internxt!"
+  },
+  {
+    "id": "halloween-campaign",
+    "title": "Internxt – Archiviazione Cloud Privata e Sicura",
+    "description": "Campagna di Halloween Internxt: Ottieni l'80% di sconto su tutti i nostri piani annuali. Proteggi i tuoi file a vita con un unico pagamento."
   }
+
 ]
