+++ src/assets/lang/it/priceCard.json
@@ -96,6 +96,15 @@
       "5TB": ["Sfondi di Star Wars", "Tema di Star Wars per Web App"],
       "10TB": ["Sfondi di Star Wars", "Tema di Star Wars per Web App", "Adesivi di Star Wars"]
     },
+    "halloweenFeatures": {
+      "title": "Gadget a tema Internxt:",
+      "200GB": ["Zaino Internxt x Swiss Peak", "Tema di Halloween per l'app Web"],
+      "1TB": ["Zaino Internxt x Swiss Peak", "Tema di Halloween per l'app Web"],
+      "2TB":  ["Zaino Internxt x Swiss Peak", "Tema di Halloween per l'app Web"],
+      "5TB":  ["Zaino Internxt x Swiss Peak", "Tema di Halloween per l'app Web"],
+      "10TB":  ["Zaino Internxt x Swiss Peak", "Tema di Halloween per l'app Web"],
+      "tooltip": "Acquista un piano a vita per avere la possibilità di vincere!"
+    },
     "cheaperThan": {
       "lifetime": {
         "normal": "Fino a ",
