+++ src/assets/lang/pt-br/banners.json
@@ -151,9 +151,9 @@
   },
   "TopBarBanner": {
     "title": {
-      "boldText": "80% de desconto em todos os planos!",
-      "normalText": "Economize no Mês de Conscientização Cibernética!",
-      "cta": "Aproveite a oferta"
+      "boldText": "80% de desconto no seu próximo plano vitalício",
+      "normalText": "Olá, quero jogar um jogo!",
+      "cta": "Você se atreve?"
     }
   },
   "lifetimeBanner": {
