+++ src/assets/lang/pt-br/metatags-descriptions.json
@@ -343,5 +343,11 @@
     "id": "image-to-text",
     "title": "Converter Imagens para Arquivos de Texto Online Gratuitamente | Internxt",
     "description": "O Conversor de Arquivos da Internxt converte suas imagens para arquivos de texto gratuitamente. Nosso conversor de imagem para texto é rápido e seguro, mantendo seus arquivos 100% privados."
+  },
+  {
+    "id": "halloween-campaign",
+    "title": "Internxt – Armazenamento em Nuvem Privado e Seguro",
+    "description": "Campanha de Halloween da Internxt: 80% de desconto em todos os nossos planos anuais. Proteja seus arquivos por toda a vida com um único pagamento."
   }
+
 ]
