+++ src/assets/lang/ru/banners.json
@@ -117,9 +117,9 @@
   },
   "TopBarBanner": {
     "title": {
-      "boldText": "Скидка 80% на все планы!",
-      "normalText": "Скидки в честь Месяца осведомленности о кибербезопасности!",
-      "cta": "Получить предложение"
+      "boldText": "Скидка 80% на ваш следующий пожизненный план",
+      "normalText": "Привет, я хочу сыграть в игру!",
+      "cta": "Осмелишься?"
     }
   },
   "AndroidSmartBanner": {
