+++ src/assets/lang/ru/home.json
@@ -65,7 +65,7 @@
       "description": {
         "normal1": "",
         "blue": "СКИДКА 80% ",
-        "normal2": "на все премиум-планы."
+        "normal2": "на все пожизненные планы"
       },
       "features": ["До 10 ТБ хранилища", "Шифрование с нулевым разглашением", "Открытый исходный код и аудит"],
       "startFrom": {
