+++ src/assets/lang/ru/metatags-descriptions.json
@@ -148,5 +148,11 @@
     "id": "cloud-storage-for-education",
     "title": "Cloud Storage Discount for Education | Internxt",
     "description": "Internxt cloud storage discount for education, schools, and universities. 50% off any cloud storage subscription plans. Keep your files secure with Internxt!"
+  },
+  {
+    "id": "halloween-campaign",
+    "title": "Internxt – Частное и безопасное облачное хранилище",
+    "description": "Хэллоуин-кампания Internxt: Скидка 80% на все наши годовые планы. Защитите свои файлы на всю жизнь с одной оплатой."
   }
+
 ]
