+++ src/assets/lang/ru/priceCard.json
@@ -96,6 +96,15 @@
       "5TB": ["Обои Звездных войн", "Тема Звездных войн для веб-приложения"],
       "10TB": ["Обои Звездных войн", "Тема Звездных войн для веб-приложения", "Наклейки Звездных войн"]
     },
+    "halloweenFeatures": {
+      "title": "Темные призы от Internxt:",
+      "200GB": ["Рюкзак Internxt x Swiss Peak", "Тема Хэллоуина для веб-приложения"],
+      "1TB": ["Рюкзак Internxt x Swiss Peak", "Тема Хэллоуина для веб-приложения"],
+      "2TB":  ["Рюкзак Internxt x Swiss Peak", "Тема Хэллоуина для веб-приложения"],
+      "5TB":  ["Рюкзак Internxt x Swiss Peak", "Тема Хэллоуина для веб-приложения"],
+      "10TB":  ["Рюкзак Internxt x Swiss Peak", "Тема Хэллоуина для веб-приложения"],
+      "tooltip": "Купите любой пожизненный план, чтобы получить шанс выиграть!"
+    },
     "cheaperThan": {
       "lifetime": {
         "normal": "До ",
