+++ src/assets/lang/ru/pricing.json
@@ -36,13 +36,17 @@
   },
   "tableSection": {
     "ctaBanner": {
-      "label": "Экономьте 80%",
-      "title": "Месяц кибербезопасности",
-      "subtitle": "Все премиум-планы Internxt!",
-      "cta": "Выбрать план",
-      "cta2": "Получите скидку 75%",
+      "label": "Сэкономьте 80 %",
+      "title": "Смертельные Хэллоуинские Скидки",
+      "subtitle": "Только на пожизненные планы",
+      "cta": "Выберите план",
+      "cta2": "Получите скидку 80 %",
       "guarantee": "30-дневная гарантия возврата денег",
-      "lastCta": "*Предложение действует для бесплатных аккаунтов или новых клиентов"
+      "lastCta": "*Предложение действительно для бесплатных аккаунтов или новых клиентов",
+      "titleMbl1": "Смертельные",
+      "titleMbl2": "скидки",
+      "titleMbl3": "на Хэллоуин.",
+      "ctaMbl": "Выбрать план"
     },
     "save": "Экономьте до",
     "planTitles": {
