+++ src/assets/lang/zh-tw/banners.json
@@ -143,9 +143,9 @@
   },
   "TopBarBanner": {
     "title": {
-      "boldText": "所有方案享 80% 折扣!",
-      "normalText": "網路安全意識月優惠！",
-      "cta": "領取優惠"
+      "boldText": "終身計劃享 8 折優惠",
+      "normalText": "你好，我想玩個遊戲！",
+      "cta": "你敢挑戰嗎？"
     }
   },
   "lifetimeBanner": {
