+++ src/assets/lang/zh-tw/home.json
@@ -65,7 +65,7 @@
       "description": {
         "normal1": "",
         "blue": "80% 折扣 ",
-        "normal2": "所有高級計劃。"
+        "normal2": "所有終身計劃"
       },
       "features": ["最高 10TB 存儲", "零知識加密", "開源和審計"],
       "startFrom": {
