+++ src/assets/lang/zh-tw/metatags-descriptions.json
@@ -318,5 +318,11 @@
     "id": "image-to-text",
     "title": "免費在線將圖像轉換為文本文件 | Internxt",
     "description": "Internxt 的文件轉換器免費將您的圖像轉換為文本文件。我們的圖像轉文本轉換器是一種快速且安全的方式，保護您的文件 100% 私密。"
+  },
+  {
+    "id": "halloween-campaign",
+    "title": "Internxt – 私密安全的雲端存儲",
+    "description": "Internxt 萬聖節活動：所有年度計劃享 8 折優惠。一次支付，終生保護您的文件。"
   }
+
 ]
