+++ src/assets/lang/zh/home.json
@@ -65,7 +65,7 @@
       "description": {
         "normal1": "",
         "blue": "80% 折扣 ",
-        "normal2": "所有高级计划。"
+        "normal2": "所有终身计划"
       },
       "features": ["最高 10TB 存储", "零知识加密", "开源和审计"],
       "startFrom": {
