+++ src/assets/lang/zh/metatags-descriptions.json
@@ -148,5 +148,11 @@
     "id": "cloud-storage-for-education",
     "title": "Cloud Storage Discount for Education | Internxt",
     "description": "Internxt cloud storage discount for education, schools, and universities. 50% off any cloud storage subscription plans. Keep your files secure with Internxt!"
+  },
+  {
+    "id": "halloween-campaign",
+    "title": "Internxt – 私密安全的云存储",
+    "description": "Internxt 万圣节活动：所有年度计划享受 8 折优惠。一次支付，终生保护您的文件。"
   }
+
 ]
